Rename laos_gggi to climate_risk
No compatibility shim: 128 of the 139 old imports in the notebooks are
submodule imports a star-import shim cannot intercept, and the archive
is read from pre-rename history.
