+++ .pre-commit-config.yaml
@@ -20,7 +20,7 @@ repos:
   rev: 0.3.0
   hooks:
     - id: no-print-statements
-      exclude: 'laos_gggi/statistics.py'
+      exclude: 'climate_risk/statistics.py'
     - id: check-execution-order
 - repo: local
   hooks:
