+++ climate_risk/data_functions/hadcrut_data_loader.py
@@ -10,8 +10,8 @@ import xarray as xr
 
 from pyprojroot import here
 
-from laos_gggi.const_vars import HADCRUT_URL
-from laos_gggi.data_functions.shapefiles_data_loader import load_shapefile
+from climate_risk.const_vars import HADCRUT_URL
+from climate_risk.data_functions.shapefiles_data_loader import load_shapefile
 
 _log = logging.getLogger(__name__)
