+++ climate_risk/data_functions/rivers_damage.py
@@ -6,15 +6,15 @@ import pandas as pd
 
 from pyprojroot import here
 
-from laos_gggi.const_vars import (
+from climate_risk.const_vars import (
     LAOS_LOCATION_DICTIONARY,
     RIVERS_FLOODS_DAMAGE_FILENAME,
     RIVERS_HYDRO_DAMAGE_FILENAME,
 )
-from laos_gggi.data_functions.emdat_processing import load_emdat_data
-from laos_gggi.data_functions.rivers_data_loader import load_rivers_data
-from laos_gggi.data_functions.shapefiles_data_loader import load_shapefile
-from laos_gggi.statistics import get_distance_to_rivers
+from climate_risk.data_functions.emdat_processing import load_emdat_data
+from climate_risk.data_functions.rivers_data_loader import load_rivers_data
+from climate_risk.data_functions.shapefiles_data_loader import load_shapefile
+from climate_risk.statistics import get_distance_to_rivers
 
 
 def create_hydro_rivers_damage():
