+++ climate_risk/plotting.py
@@ -12,8 +12,8 @@ from matplotlib.gridspec import GridSpec
 from matplotlib.offsetbox import AnchoredText
 from scipy import stats
 
-from laos_gggi.const_vars import REGIONS
-from laos_gggi.model import get_distance_to
+from climate_risk.const_vars import REGIONS
+from climate_risk.model import get_distance_to
 
 
 def configure_plot_style(add_grid=False):
