--- laos_gggi/data_functions/__init__.py
@@ -1,29 +1,0 @@
-from laos_gggi.data_functions.co2_processing import load_co2_data
-from laos_gggi.data_functions.combine_data import load_all_data
-from laos_gggi.data_functions.disaster_point_data import (
-    load_disaster_point_data,
-    load_synthetic_non_disaster_points,
-)
-from laos_gggi.data_functions.emdat_processing import load_emdat_data
-from laos_gggi.data_functions.GPCC_data_loader import load_gpcc_data
-from laos_gggi.data_functions.hadcrut_data_loader import load_hadcrut_data
-from laos_gggi.data_functions.ipcc_scenarios_loader import process_ipcc_scenarios
-from laos_gggi.data_functions.ocean_heat_processing import load_ocean_heat_data
-from laos_gggi.data_functions.rivers_data_loader import load_rivers_data
-from laos_gggi.data_functions.shapefiles_data_loader import load_shapefile
-from laos_gggi.data_functions.world_bank_data_loader import load_wb_data
-
-__all__ = [
-    "load_all_data",
-    "load_co2_data",
-    "load_disaster_point_data",
-    "load_emdat_data",
-    "load_gpcc_data",
-    "load_hadcrut_data",
-    "load_ocean_heat_data",
-    "load_rivers_data",
-    "load_shapefile",
-    "load_synthetic_non_disaster_points",
-    "load_wb_data",
-    "process_ipcc_scenarios",
-]
